New Changes

+++ lib/farmerSide/views/item_screen.dart
@@ -1,7 +1,8 @@
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
-
+import 'package:harvest_delivery/main.dart';
+import 'package:firebase_auth/firebase_auth.dart';
 import 'add_item.dart';
 import 'update_item.dart';
 
@@ -14,22 +15,47 @@ class ItemScreen extends StatefulWidget {
 }
 
 class _ItemScreenState extends State<ItemScreen> {
-  final Stream<QuerySnapshot> itemsStream =
-      FirebaseFirestore.instance.collection('MarketProducts').snapshots();
+    final Stream<QuerySnapshot> itemsStream = FirebaseFirestore.instance
+      .collection('MarketProducts')
+      .where('FarmerId', isEqualTo: FirebaseAuth.instance.currentUser!.uid)
+      .snapshots();
 
-  Future<void> deleteItem(String id) async {
-    try {
+Future<void> deleteItem(String id) async {
+  try {
+    bool confirmDelete = await showDialog(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: Text('Confirm Deletion'),
+          content: Text('Are you sure you want to delete this item?'),
+          actions: [
+            TextButton(
+              onPressed: () => Navigator.of(context).pop(false),
+              child: Text('Cancel'),
+            ),
+            TextButton(
+              onPressed: () => Navigator.of(context).pop(true),
+              child: Text('Delete'),
+            ),
+          ],
+        );
+      },
+    );
+
+    if (confirmDelete == true) {
       await FirebaseFirestore.instance
           .collection('MarketProducts')
           .doc(id)
           .delete();
-    } catch (error) {
-      print('Failed to delete item: $error');
-      ScaffoldMessenger.of(context).showSnackBar(
-        SnackBar(content: Text('Failed to delete item')),
-      );
     }
+  } catch (error) {
+    print('Failed to delete item: $error');
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(content: Text('Failed to delete item')),
+    );
   }
+}
+
 
   Future<String> getImageURL(String imagePath) async {
     final imageRef = FirebaseStorage.instance.ref().child(imagePath);
@@ -139,7 +165,7 @@ class _ItemScreenState extends State<ItemScreen> {
             MaterialPageRoute(builder: (context) => AddItem()),
           );
         },
-        backgroundColor: const Color.fromARGB(255, 243, 159, 33),
+        backgroundColor: MyApp.primaryColor,
         foregroundColor: Colors.white,
         child: Icon(Icons.add),
       ),
